feat: Update types #201
